requirements.txt
# Core MCP and Web Framework
mcp[server-fastapi]==1.1.2
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Image Processing
Pillow==10.4.0
face-recognition==1.3.0
dlib==19.24.6

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1

# AI Integration
openai==1.54.3
anthropic==0.39.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.4
